+++ project-scripts/utils.js
@@ -20,7 +20,10 @@ const compilerPromise = (name, compiler) => {
   });
 };
 
+const sleep = ms => new Promise(resolve => setTimeout(resolve, ms));
+
 module.exports = {
   logMessage,
-  compilerPromise
+  compilerPromise,
+  sleep,
 };