+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 ## 0.1.1
 
 * Installed major packages.
+* Set up linting and typechecking.
 * Built a changelog and readme.
 * Webpack config files for production and development.
+* Webpack config files for client and server.
 * Added CSS reset.
+* Added server-side rendering.
+* Created cra-esq project-scripts for use in package.json.
+* Added hot module replacement.
+* Added redux store.
 
 ## 0.1.0